+++ spec/dicey/cli/calculator_runner_spec.rb
@@ -1,12 +1,14 @@
 # frozen_string_literal: true
 
+require "dicey/cli/calculator_runner"
+
 module Dicey
-  RSpec.describe SumFrequencyCalculators::Runner do
+  RSpec.describe CLI::CalculatorRunner do
     subject(:call_result) { described_class.new.call(dice, format: format, result: result) }
 
     let(:dice) { %w[2d2 1,5,] } # rubocop:disable Lint/PercentStringArray
     let(:calculators) { [SumFrequencyCalculators::KroneckerSubstitution.new] }
-    let(:format) { OutputFormatters::JSONFormatter.new }
+    let(:format) { CLI::Formatters::JSONFormatter.new }
     let(:result) { :frequencies }
 
     before { stub_const("Dicey::SumFrequencyCalculators::AutoSelector::AVAILABLE_CALCULATORS", calculators) }