+++ spec/dicey/cli/roller_spec.rb
@@ -1,11 +1,13 @@
 # frozen_string_literal: true
 
+require "dicey/cli/roller"
+
 module Dicey
-  RSpec.describe Roller do
+  RSpec.describe CLI::Roller do
     subject(:result) { described_class.new.call(dice, format: format) }
 
     let(:dice) { %w[2 (1,5) 2D3] }
-    let(:format) { OutputFormatters::ListFormatter.new }
+    let(:format) { CLI::Formatters::ListFormatter.new }
 
     it "returns a formatted string with dice description and roll result" do
       expect(result).to match(/\A# D2\+\(1,5\)\+D3\+D3\nroll => (?:[4-9]|1[0-3])\n\z/)